Check for environment and run the correct thing;

+++ launch.sh
@@ -15,4 +15,9 @@ theme=$(echo $theme | tr -d '"')
 cp -r /app/themes/$theme/* /app
 npm install -g npm@latest
 npm install
-npm run build && rm /app/.env && npm start
+
+if [ "$ENV" = "development" ]; then
+  npm run dev
+else
+  npm run build && rm /app/.env && npm start
+fi
